# specs/004-production-hardening/contracts/api.yaml
openapi: 3.0.0
info:
  title: Production Hardening API
  version: 1.0.0
  description: API contracts for production hardening, monitoring, and error handling
paths:
...
  /api/monitoring/errors:
    post:
      summary: Report application error
      description: Submit error report for tracking and analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrorReport'
      responses:
        '201':
          description: Error reported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [logged, escalated]
        '400':
          description: Invalid error report format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorReport:
      type: object
      properties:
        boundaryContext:
          type: object
          properties:
            id:
              type: string
              format: uuid
            component:
              type: string
            errorType:
              type: string
              enum: [ai_service, network, validation, unknown]
            timestamp:
              type: string
              format: date-time
            userAgent:
              type: string
            routePath:
              type: string
            metadata:
              type: object
          required: [id, component, errorType, timestamp, routePath]
        error:
          type: object
          properties:
            name:
              type: string
            message:
              type: string
            stack:
              type: string
            cause:
              type: object
          required: [name, message]
        recovery:
          type: object
          properties:
            attempted:
              type: boolean
            successful:
              type: boolean
            fallbackUsed:
              type: string
          required: [attempted, successful]
        user:
          type: object
          properties:
            sessionId:
              type: string
            actionHistory:
              type: array
              items:
                type: string
          required: [sessionId, actionHistory]
      required: [boundaryContext, error, recovery, user]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
      required: [error, message, timestamp]